// src/app/models/role.enum.ts
export enum RoleEnum {
  ROLE_USER = "ROLE_USER",
  ROLE_SHD = "ROLE_SHD",
  ROLE_BADGE = "ROLE_BADGE",
  ROLE_RH = "ROLE_RH",
  ROLE_IT = "ROLE_IT",
  ROLE_SURFACE = "ROLE_SURFACE",
  ROLE_ORDER = "ROLE_ORDER",
  ROLE_MIL = "ROLE_MIL",
  ROLE_ASSISTANT = "ROLE_ASSISTANT",
  ROLE_ADMIN = "ROLE_ADMIN",
}